// api/src/cnpj/entities/socio.entity.ts
export interface SocioDocument {
  cnpj_basico: string;
  identificador_de_socio?: string;
  nome_socio: string;
  cnpj_cpf_do_socio?: string;
  qualificacao_socio?: string;
  data_entrada_sociedade?: string;
  pais?: string;
  representante_legal?: string;
  nome_do_representante?: string;
  qualificacao_representante_legal?: string;
  faixa_etaria?: string;
}